refactor(settings): website/settings.py をディレクトリ化 (#332)
548行に各種設定が混在していた settings.py を責務別の6ファイルに分割した。

- settings/__init__.py: パッケージのロード順序を定義
- settings/base.py: Django コア（INSTALLED_APPS / MIDDLEWARE / TEMPLATES /
  DATABASES / CACHES / LOGGING / I18N / STATIC / AUTH_PASSWORD_VALIDATORS）
- settings/security.py: ALLOWED_HOSTS / CSRF / CORS / HSTS /
  SECURE_PROXY_SSL_HEADER / DISCORD_AUTH_REQUIRED
- settings/apis.py: Google / Gemini / GA4 / Amazon SES / Discord Webhook
- settings/storage.py: Cloudflare R2 / MEDIA / STORAGES
- settings/caching.py: 将来の Redis/Session 拡張用スケルトン
- settings/authentication.py: AUTH_USER_MODEL / allauth / Discord OAuth

ロード順は __init__.py で base → security → apis → storage → caching →
authentication に固定。DEBUG/BASE_DIR/TESTING/共通ヘルパー (_mask /
_split_csv_env / _settings_logger) は base.py に集約し、他モジュールは
`from .base import ...` で取り込む。

後方互換性: `from website.settings import X` の既存 import (12ファイル)
は無変更で動作。直接 `website.settings._build_allowed_hosts()` を呼んで
いた test_host_settings.py のみ `from website.settings.security import
_build_allowed_hosts` に書き換え。

検証:
- docker compose exec vrc-ta-hub python manage.py check OK
- 全1600 tests OK (skipped=13)
- docker compose restart 後 http://localhost:8015/ で HTTP 200 OK
- トップページのスクリーンショットで UI 崩れなしを確認

Closes #332
