+++ app/website/settings/base.py
@@ -1,20 +1,13 @@
-"""
-Django settings for website project.
+"""Django コア設定.
 
-Generated by 'django-admin startproject' using Django 4.2.4.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.2/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/4.2/ref/settings/
+INSTALLED_APPS / MIDDLEWARE / TEMPLATES / DATABASES / CACHES / LOGGING / I18N /
+STATIC など、Django プロジェクトの骨格を定義する。BASE_DIR / SECRET_KEY / DEBUG /
+TESTING もここで定義し、他の設定モジュールは `from .base import ...` で参照する。
 """
 import logging
 import os
 import sys
 from pathlib import Path
-
-from website.hosts import get_canonical_host, normalize_host
 
 _settings_logger = logging.getLogger('django.settings')
 
@@ -32,30 +25,8 @@ def _split_csv_env(env_name: str) -> list[str]:
     return [item.strip() for item in value.split(',') if item.strip()]
 
 
-def _get_canonical_host() -> str:
-    return get_canonical_host()
-
-
-APP_CANONICAL_HOST = _get_canonical_host()
-
-
-def _build_allowed_hosts() -> list[str]:
-    hosts = [
-        _get_canonical_host(),
-        'localhost',
-        '127.0.0.1',
-        *_split_csv_env('ALLOWED_HOSTS'),
-    ]
-
-    http_host = normalize_host(os.environ.get('HTTP_HOST'))
-    if http_host:
-        hosts.append(http_host)
-
-    return list(dict.fromkeys(hosts))
-
-
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
-BASE_DIR = Path(__file__).resolve().parent.parent
+BASE_DIR = Path(__file__).resolve().parent.parent.parent
 
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/4.2/howto/deployment/checklist/
@@ -66,26 +37,6 @@ SECRET_KEY = os.environ.get('SECRET_KEY')
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = os.environ.get('DEBUG') == 'True'
 
-ALLOWED_HOSTS = _build_allowed_hosts()
-
-# Cloud Run + nginx プロキシ経由の HTTPS 判定（本番: nginx が https を付加）
-# ローカルでは .env.local で HTTP_X_FORWARDED_PROTO=http を設定して is_secure()=False を保証する。
-SECURE_PROXY_SSL_HEADER = ('HTTP_X_FORWARDED_PROTO', 'https')
-
-# 本番環境のセキュリティ強化
-if not DEBUG:
-    SESSION_COOKIE_SECURE = True
-    CSRF_COOKIE_SECURE = True
-    SECURE_HSTS_SECONDS = 31536000  # 1年
-    SECURE_HSTS_INCLUDE_SUBDOMAINS = True
-    SECURE_HSTS_PRELOAD = True
-
-CSRF_TRUSTED_ORIGINS = [
-    o for o in [
-        'https://vrc-ta-hub.com',
-        os.environ.get('CSRF_TRUSTED_ORIGIN'),
-    ] if o
-]
 # Application definition
 
 INSTALLED_APPS = [
@@ -149,11 +100,6 @@ SPECTACULAR_SETTINGS = {
     ],
 }
 
-CORS_ALLOW_ALL_ORIGINS = True
-CORS_URLS_REGEX = r'^/api/.*$'
-
-AUTH_USER_MODEL = 'user_account.CustomUser'
-
 MIDDLEWARE = [
     # Cloud Run preview host は raw Host のまま下流へ流さない。
     'website.middleware.CanonicalCloudRunHostMiddleware',
@@ -167,11 +113,6 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
     'allauth.account.middleware.AccountMiddleware',
     'user_account.middleware.DiscordAuthRequiredMiddleware',
-]
-
-AUTHENTICATION_BACKENDS = [
-    'django.contrib.auth.backends.ModelBackend',
-    'allauth.account.auth_backends.AuthenticationBackend',
 ]
 
 ROOT_URLCONF = 'website.urls'
@@ -212,11 +153,6 @@ DATABASES = {
     },
 }
 
-# Discord認証強制ミドルウェアの設定
-# DEBUG=True（開発環境）では無効化（ブラウザ操作MCPでのテストのため）
-# 本番環境（DEBUG=False）では有効化
-DISCORD_AUTH_REQUIRED = not DEBUG
-
 TESTING = os.environ.get('TESTING', '').strip().lower() in {'1', 'true', 'yes', 'on'}
 
 if 'test' in sys.argv or TESTING:
@@ -224,9 +160,6 @@ if 'test' in sys.argv or TESTING:
         'ENGINE': 'django.db.backends.sqlite3',
         'NAME': 'test_db.sqlite3',
     }
-    # テスト環境ではDiscord認証強制ミドルウェアを無効化
-    # 個別のテストで検証する場合は override_settings で有効化
-    DISCORD_AUTH_REQUIRED = False
 
 _settings_logger.info('DB_NAME: %s', _mask(DATABASES['default']['NAME']))
 
@@ -286,106 +219,6 @@ FORMS_URLFIELD_ASSUME_HTTPS = True
 
 LOGIN_URL = '/account/login/'
 LOGIN_REDIRECT_URL = '/event/my_list/'
-
-# Google Calendar APIの設定
-GOOGLE_CALENDAR_CREDENTIALS = os.getenv('GOOGLE_CALENDAR_CREDENTIALS', '/app/credentials.json')
-
-# Cloudflare R2の設定
-AWS_ACCESS_KEY_ID = os.getenv('AWS_ACCESS_KEY_ID')
-AWS_SECRET_ACCESS_KEY = os.getenv('AWS_SECRET_ACCESS_KEY')
-AWS_STORAGE_BUCKET_NAME = os.getenv('AWS_STORAGE_BUCKET_NAME')
-AWS_S3_ENDPOINT_URL = os.getenv('AWS_S3_ENDPOINT_URL')
-AWS_S3_CUSTOM_DOMAIN = os.getenv('AWS_S3_CUSTOM_DOMAIN')
-AWS_S3_URL_PROTOCOL = os.getenv('AWS_S3_URL_PROTOCOL', 'https:')
-AWS_S3_SECURE_URLS = AWS_S3_URL_PROTOCOL == 'https:'
-
-# ファイルストレージ（メディア）
-if AWS_STORAGE_BUCKET_NAME:
-    # R2を使用（本番・開発環境）
-    MEDIA_URL = f'{AWS_S3_ENDPOINT_URL}/{AWS_STORAGE_BUCKET_NAME}/'
-    AWS_S3_FILE_OVERWRITE = False
-    AWS_QUERYSTRING_AUTH = False  # 認証付きのURLを生成しない
-    default_storage = {
-        'BACKEND': 'storages.backends.s3boto3.S3Boto3Storage',
-        'OPTIONS': {
-            'access_key': AWS_ACCESS_KEY_ID,
-            'secret_key': AWS_SECRET_ACCESS_KEY,
-            'bucket_name': AWS_STORAGE_BUCKET_NAME,
-            'endpoint_url': AWS_S3_ENDPOINT_URL,
-            'custom_domain': AWS_S3_CUSTOM_DOMAIN,
-            'file_overwrite': AWS_S3_FILE_OVERWRITE,
-            'querystring_auth': AWS_QUERYSTRING_AUTH,
-        },
-    }
-else:
-    # ローカルファイルストレージを使用（テスト環境）
-    MEDIA_URL = '/media/'
-    MEDIA_ROOT = BASE_DIR / 'media'
-    default_storage = {
-        'BACKEND': 'django.core.files.storage.FileSystemStorage',
-    }
-
-if DEBUG:
-    # 静的ファイルはローカル配信（開発時の利便性）
-    staticfiles_storage = {
-        'BACKEND': 'django.contrib.staticfiles.storage.StaticFilesStorage',
-    }
-else:
-    # 本番は静的ファイルもR2
-    staticfiles_storage = {
-        'BACKEND': 'storages.backends.s3boto3.S3Boto3Storage',
-        'OPTIONS': {
-            'access_key': AWS_ACCESS_KEY_ID,
-            'secret_key': AWS_SECRET_ACCESS_KEY,
-            'bucket_name': AWS_STORAGE_BUCKET_NAME,
-            'endpoint_url': AWS_S3_ENDPOINT_URL,
-            'custom_domain': AWS_S3_CUSTOM_DOMAIN,
-        },
-    }
-
-STORAGES = {
-    'default': default_storage,
-    'staticfiles': staticfiles_storage,
-}
-
-# Google API
-GOOGLE_API_KEY = os.environ.get('GOOGLE_API_KEY')
-assert GOOGLE_API_KEY is not None, 'Please set GOOGLE_API_KEY'
-GOOGLE_CALENDAR_ID = os.environ.get('GOOGLE_CALENDAR_ID')
-assert GOOGLE_CALENDAR_ID is not None, 'Please set GOOGLE_CALENDAR_ID'
-_settings_logger.info('GOOGLE_CALENDAR_ID: %s', _mask(GOOGLE_CALENDAR_ID))
-if GOOGLE_CALENDAR_ID.startswith('d80eac'):
-    _settings_logger.info('Debug mode: GOOGLE_CALENDAR_ID starts with d80eac')
-else:
-    _settings_logger.info('Production mode: GOOGLE_CALENDAR_ID')
-GEMINI_API_KEY = os.environ.get('GEMINI_API_KEY')
-assert GEMINI_API_KEY is not None, 'Please set GEMINI_API_KEY'
-GEMINI_MODEL = os.environ.get('GEMINI_MODEL', 'google/gemini-2.5-flash-lite-preview-06-17')
-_settings_logger.info('GEMINI_MODEL: %s', GEMINI_MODEL)
-
-# GA4 Data API（ページ別アクセス解析）
-# GA4_PROPERTY_ID は Data API 用の数値ID。base.html の measurement ID G-6BN9EHVMRW とは別物
-GA4_PROPERTY_ID = os.environ.get('GA4_PROPERTY_ID', '444114283')
-GOOGLE_APPLICATION_CREDENTIALS = os.getenv(
-    'GOOGLE_APPLICATION_CREDENTIALS', '/app/secret/credentials.json'
-)
-
-REQUEST_TOKEN = os.environ.get('REQUEST_TOKEN')
-assert REQUEST_TOKEN is not None, 'Please set REQUEST_TOKEN'
-
-# メール設定 (Amazon SES)
-if DEBUG:
-    EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'
-else:
-    EMAIL_BACKEND = 'django_ses.SESBackend'
-AWS_SES_REGION_NAME = os.environ.get('AWS_SES_REGION_NAME', 'ap-northeast-1')
-AWS_SES_REGION_ENDPOINT = os.environ.get('AWS_SES_REGION_ENDPOINT', 'email.ap-northeast-1.amazonaws.com')
-AWS_SES_ACCESS_KEY_ID = os.environ.get('AWS_SES_ACCESS_KEY_ID')  # SES専用のアクセスキー
-AWS_SES_SECRET_ACCESS_KEY = os.environ.get('AWS_SES_SECRET_ACCESS_KEY')  # SES専用のシークレットキー
-DEFAULT_FROM_EMAIL = os.environ.get('DEFAULT_FROM_EMAIL', 'VRC技術学術系Hub <info@vrc-ta-hub.com>')
-
-# Admin email for notifications
-ADMIN_EMAIL = os.environ.get('ADMIN_EMAIL', DEFAULT_FROM_EMAIL)
 
 # ログディレクトリの設定
 LOG_DIR = BASE_DIR / 'logs'
@@ -500,49 +333,3 @@ if DEBUG:
     for logger_name in LOGGING['loggers']:
         LOGGING['loggers'][logger_name]['handlers'].append('file')
         LOGGING['loggers'][logger_name]['level'] = 'DEBUG'
-
-# django-allauth 設定
-ACCOUNT_EMAIL_REQUIRED = False
-ACCOUNT_USERNAME_REQUIRED = False
-ACCOUNT_AUTHENTICATION_METHOD = 'username'
-ACCOUNT_EMAIL_VERIFICATION = 'none'
-ACCOUNT_USER_MODEL_USERNAME_FIELD = 'user_name'
-ACCOUNT_SESSION_REMEMBER = None  # ユーザーに選択させる
-SOCIALACCOUNT_AUTO_SIGNUP = True
-SOCIALACCOUNT_LOGIN_ON_GET = True
-SOCIALACCOUNT_ADAPTER = 'user_account.adapters.CustomSocialAccountAdapter'
-
-# OAuth callback URLのプロトコル（本番: https、開発: http）
-ACCOUNT_DEFAULT_HTTP_PROTOCOL = os.environ.get('ACCOUNT_DEFAULT_HTTP_PROTOCOL', 'https')
-
-# Discord OAuth設定
-DISCORD_CLIENT_ID = os.environ.get('DISCORD_CLIENT_ID', '')
-DISCORD_CLIENT_SECRET = os.environ.get('DISCORD_CLIENT_SECRET', '')
-
-SOCIALACCOUNT_PROVIDERS = {
-    'discord': {
-        'SCOPE': ['identify', 'email'],
-    }
-}
-
-# 環境変数が設定されている場合のみAPPS設定を追加
-if DISCORD_CLIENT_ID and DISCORD_CLIENT_SECRET:
-    SOCIALACCOUNT_PROVIDERS['discord']['APPS'] = [
-        {
-            'client_id': DISCORD_CLIENT_ID,
-            'secret': DISCORD_CLIENT_SECRET,
-            'key': '',
-        }
-    ]
-
-SOCIALACCOUNT_FORMS = {
-    'signup': 'user_account.forms.CustomSocialSignupForm',
-}
-
-# ソーシャルアカウントの接続解除（disconnect）を試みた場合のリダイレクト先
-# 削除ボタンはテンプレートで非表示にするが、直接アクセスされた場合の保険
-SOCIALACCOUNT_DISCONNECT_REDIRECT_URL = '/account/settings/'
-
-# Discord Webhook（管理者通知用）
-DISCORD_WEBHOOK_URL = os.environ.get('DISCORD_WEBHOOK_URL', '')
-DISCORD_REPORT_WEBHOOK_URL = os.environ.get('DISCORD_REPORT_WEBHOOK_URL', '')
